Update Ship::Launch()
Health checks correspond to the current health checks that cause a fighter to board.
Fighters will not wait to be repaired if the parent will not be able to repair that particular aspect soon.
Ramscoop collection first collects the fuel for a jump, then refuels available fighters, then resumes restocking the parent's reserves. This provides visual indication of fuel availability for the player.
Fighters that board will empty their fuel tanks into the parent's supply.
Fighters that depart will attempt to top off if needed.
If a fighter can fight without ammo and/or fuel, it will not wait to launch until getting them.
Added checks to ensure fighters can launch if the carrier is gravely threatened.